Add comments, clean code and improve code clarity
Modified getAlphaColor() function to return 1 if color value is
of the form 'rgb(x,x,x)' since this implies alpha is 1 or 100%.

Modified logic in mouse events according to the updated function.

Extracted duplicate code from mouse events into a function called
colorGridItem(gridItem) which takes a grid item and colors it. The mouse
events will call this function instead.

Added some comments to explain the code for future reference.